// libs/features/occupational-hazard/src/lib/shared/enums/allowance-type.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum AllowanceType {
  INPATIENT = 'Inpatient Allowance',
  OUTPATIENT = 'OutPatient Allowance',
  BALANCE_SETTLEMENT = 'Balance Settlement Allowance',
  BALANCE_SETTLEMENT_REVERSAL = 'Balance Settlement Reversal',
  CONV_ALLOWANCE = 'Conveyance Allowance',
  COMP_CONV_ALLOWANCE = 'Companion Conveyance Allowance',
  COMP_ALLOWANCE = 'Companion Allowance',
  DAILY_ALLOWANCE = 'Daily Allowance',
  TOTAL_DIS_REPATRIATION = 'Total Disability Repatriation',
  DEAD_BODY_REPATRIATION = 'Dead Body Repatriation',
  CONTRIBUTOR = 2,
  ESTABLISHMENT = 1
}