// libs/features/occupational-hazard/src/lib/shared/models/allowance-filter-params.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export class AllowanceFilterParams {
  claimType: string[] = undefined;
  noOfDaysMax: number;
  noOfDaysMin: number;
  visitsMin: number;
  visitsMax: number;
}